Refactor models/auto-models.py and add a load of command line arguments
+ Command line arguments allow different model parameters, such as
globule radius, ionizing flux, etc.

+ Hybrid pivoting/bisection algorithm is improved (but still isn't
watertight)

+ The actual running of the Cloudy model is now encapsulated as a
function. This is in preparation for switching to a more sophisticated
root-finding algorithm, such as scipy.optimize.brentq. The shared
state that is needed is held in just two variables: "args", which
holds the command-line arguments, and "anglepars", which holds the
additional parameters that are specific to this particular angle of
incidence .