Improve the root-finding method in auto-models.py
Just use secant method to bracket the roots, then start again with brent method. This should be guaranteed to work in all circumstances, and initial tests suggest it will converge in fewer iterations. We also cache the function evaluations, so that we never have to run the same Cloudy model twice.