New argument --read-extra-cloudy-commands-from-file for auto-models.py
This will look for a file cloudy-extras.in in the input folder.  If it
exists, it will append the contents to the end of every input script.
This is intended to allow small temporary modifications to the cloudy
models without having to modify the python programs.  However, any
modification that we want to keep should be put in "properly" in
cloudycontrol.  The argument is deliberately very verbose, and is
turned off by default, so that we will never forget that we are using
it.