The last version had problems, now works but the results are not exactly that we expected
The last version does not work because the X Y length was not the same of the mask,
this was because I used wrong the np.arenge. With np.linspace is better and easier.

In other hand, there was a switch in the rotated coordenates. With the previous
version was enought to solve it.

Now we have ALMOST the same results that we obtain with extract_aperture.py
the variation in the total fluxes are more or less 10 percent.